🎨 feat: .상품 상세 정보를 path경로값에 따라 바뀌는 기능 구현 ♻️ refactor: 컴포넌트 /  라우트 폴더 구조 변경
