+++ src/app/(user)/my_page/(account)/layout.tsx
@@ -1,0 +1,8 @@
+'use client';
+export default function MyPage({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return <main className="bg-white pb-40">{children}</main>;
+}
