+++ src/app/(user)/my_page/(account)/regiester/page.tsx
@@ -23,7 +23,7 @@ export default function Register() {
   useEffect(() => {
     if (state?.ok) {
       alert('회원 가입이 완료되었습니다. 로그인 페이지로 이동합니다.');
-      navigation.replace('/my_page/login');
+      navigation.replace('/my_page');
     } else if (state?.ok === 0 && !state?.errors) {
       // 입력값 검증에러가 아닌 경우
       alert(state?.message);
