+++ src/data/actions/products.ts
@@ -28,9 +28,9 @@ export async function getProductList(): ApiResPromise<ProductList> {
  * 상품의 상세정보를 가져옵니다.
  * @returns {Promise<ApiRes<Product[]>>} - 상품 목록 응답 객체
  */
-export async function getProductInfo(): ApiResPromise<Product> {
+export async function getProductInfo(path: string): ApiResPromise<Product> {
   try {
-    const res = await axios.get(`${API_URL}/products/2`, {
+    const res = await axios.get(`${API_URL}/products/${path}`, {
       headers: {
         'Content-Type': 'application/json',
         'Client-Id': CLIENT_ID,
